Add formatters for Unlocode and Country and use them on CYA
