+++ src/utils/Formatters/formatCountry.js
@@ -1,0 +1,9 @@
+import { countries } from '../../constants/CountryData';
+
+const formatCountry = (code) => {
+  const result = countries.find((country) => country.alphaCode === code);
+  const countryName = result?.countryName ? result.countryName : code;
+  return countryName;
+};
+
+export default formatCountry;
